Add a datasource and viewmodel
