+++ WooCommerce/Classes/ViewRelated/Notifications/ReviewsDataSource.swift
@@ -1,0 +1,130 @@
+import Foundation
+import UIKit
+import Yosemite
+
+
+final class ReviewsDataSource: NSObject {
+    lazy var resultsController: ResultsController<StorageProductReview> = {
+        let storageManager = ServiceLocator.storageManager
+        let descriptor = NSSortDescriptor(keyPath: \StorageProductReview.dateCreated, ascending: false)
+
+        return ResultsController<StorageProductReview>(storageManager: storageManager,
+                                                       sectionNameKeyPath: "normalizedAgeAsString",
+                                                       matching: self.filterPredicate,
+                                                       sortedBy: [descriptor])
+    }()
+
+    private lazy var filterPredicate: NSPredicate = {
+        let sitePredicate = NSPredicate(format: "siteID == %lld",
+                                        ServiceLocator.stores.sessionManager.defaultStoreID ?? Int.min)
+
+        let statusPredicate = NSPredicate(format: "statusKey ==[c] %@ OR statusKey ==[c] %@",
+                                          ProductReviewStatus.approved.rawValue,
+                                          ProductReviewStatus.hold.rawValue)
+
+        return  NSCompoundPredicate(andPredicateWithSubpredicates: [sitePredicate, statusPredicate])
+    }()
+
+    /// Keep track of the (Autosizing Cell's) Height. This helps us prevent UI flickers, due to sizing recalculations.
+    ///
+    private var estimatedRowHeights = [IndexPath: CGFloat]()
+}
+
+
+// MARK: - UITableViewDataSource Conformance
+//
+extension ReviewsDataSource: UITableViewDataSource {
+
+    func numberOfSections(in tableView: UITableView) -> Int {
+        return resultsController.sections.count
+    }
+
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return resultsController.sections[section].numberOfObjects
+    }
+
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        guard let cell = tableView.dequeueReusableCell(withIdentifier: NoteTableViewCell.reuseIdentifier) as? NoteTableViewCell else {
+            fatalError()
+        }
+
+        configure(cell, at: indexPath)
+
+        return cell
+    }
+
+    func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
+        let rawAge = resultsController.sections[section].name
+        return Age(rawValue: rawAge)?.description
+    }
+}
+
+
+// MARK: - Cell Setup
+//
+private extension ReviewsDataSource {
+
+    /// Initializes the Notifications Cell at the specified indexPath
+    ///
+    func configure(_ cell: NoteTableViewCell, at indexPath: IndexPath) {
+        let review = resultsController.object(at: indexPath)
+
+        let viewModel = ReviewViewModel(review: review)
+        cell.configure(with: viewModel)
+    }
+}
+
+
+extension ReviewsDataSource: UITableViewDelegate {
+
+    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        return UITableView.automaticDimension
+    }
+
+    func tableView(_ tableView: UITableView, estimatedHeightForRowAt indexPath: IndexPath) -> CGFloat {
+        return estimatedRowHeights[indexPath] ?? Settings.estimatedRowHeight
+    }
+
+    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        return UITableView.automaticDimension
+    }
+
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        tableView.deselectRow(at: indexPath, animated: true)
+
+        let review = resultsController.object(at: indexPath)
+        presentDetails(for: review)
+    }
+
+    func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
+
+        // Preserve the Cell Height
+        // Why: Because Autosizing Cells, upon reload, will need to be laid yout yet again. This might cause
+        // UI glitches / unwanted animations. By preserving it, *then* the estimated will be extremely close to
+        // the actual value. AKA no flicker!
+        //
+        estimatedRowHeights[indexPath] = cell.frame.height
+    }
+}
+
+// MARK: - Public Methods
+//
+private extension ReviewsDataSource {
+
+    /// Presents the Details for a given Note Instance: Either NotificationDetails, or OrderDetails, depending on the
+    /// Notification's Kind.
+    ///
+    func presentDetails(for review: ProductReview) {
+        print("==== presenting detils for review")
+
+        //        let detailsViewController = NotificationDetailsViewController(note: note)
+        //        navigationController?.pushViewController(detailsViewController, animated: true)
+    }
+}
+
+
+private extension ReviewsDataSource {
+    enum Settings {
+        static let estimatedRowHeight = CGFloat(88)
+    }
+}
